- delete dead code - reorganize project structure - add small config changes
