+++ .editorconfig
@@ -7,13 +7,11 @@ root = true
 [*]
 end_of_line = lf
 insert_final_newline = true
+trim_trailing_whitespace = true
 
 # Matches multiple files with brace expansion notation
 # Set default charset
 [*.{js,ts,tsx}]
-charset = utf-8
-
-# Indentation override for all JS under lib directory
-[**.js]
 indent_style = space
 indent_size = 2
+charset = utf-8
