+++ .gitignore
@@ -18,6 +18,8 @@
 .env.test.local
 .env.production.local
 
+/src/react-app-env.d.ts
+
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
