--- .nvmrc
@@ -1,1 +1,0 @@
-v10.15.3
