--- src/App.test.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-
-it('creates root node, A, with null values', () => {
-  const data = 0
-  // const div = document.createElement('div');
-  // ReactDOM.render(<App width={800} height={600} graph={data} />, div);
-  // ReactDOM.unmountComponentAtNode(div);
-});
