--- src/components/App.css
@@ -1,48 +1,0 @@
-body {
-  background-color: rgb(28,32,27);
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-header {
-  text-align: left;
-}
-
-.App-footer {
-  text-align: right;
-}
-
-.App-header pre p {
-  line-height: 0.05em;
-  font-size: 1.5em;
-}
-
-.App-footer pre p {
-  line-height: 0.05em;
-  font-size: 0.5em;
-  color: gray;
-}
-
-.viz-mount {
-  border: 1px solid crimson;
-}
-
-.node {
-  stroke: #fff;
-  stroke-width: 1.5px;
-}
-
-.link {
-  stroke: #999;
-  stroke-opacity: 1;
-}
-
-.label {
-  font-size: 10px;
-  fill: #fff;
-}
