+++ src/components/Nodes.tsx
@@ -35,39 +35,6 @@ const Nodes: React.FC<INodesProps> = (props) => {
         return <Node key={index} node={node} color={color(node.group.toString())}/>
     })
 
-    // TODO: fix d3.drag call
-    // useEffect(() => {
-    //     const simulation = props.simulation
-
-        // function onDragStart(d: any) {
-        //     if (!d3.event.active) {
-        //         simulation.alphaTarget(0.3).restart()
-        //     }
-        //     d.fx = d.x
-        //     d.fy = d.y
-        // }
-        //
-        // function onDrag(d: any) {
-        //     d.fx = d3.event.x
-        //     d.fy = d3.event.y
-        // }
-        //
-        // function onDragEnd(d: any) {
-        //     if (!d3.event.active) {
-        //         simulation.alphaTarget(0)
-        //     }
-        //     d.fx = null
-        //     d.fy = null
-        // }
-
-        // d3.selectAll(".node")
-        //     .call(
-        //         d3.drag()
-        //             .on("start", onDragStart)
-        //             .on("drag", onDrag)
-        //             .on("end", onDragEnd))
-    // })
-
     return (
         <g className="nodes">
             {nodes}
